+++ src/main/kotlin/com/example/tqi_Kotlin_backend_developer_2023/service/IShoppingCartService.kt
@@ -3,6 +3,7 @@ package com.example.tqi_Kotlin_backend_developer_2023.service
 import com.example.tqi_Kotlin_backend_developer_2023.domain.PaymentOptions
 import com.example.tqi_Kotlin_backend_developer_2023.domain.Sale
 import com.example.tqi_Kotlin_backend_developer_2023.domain.ShoppingCart
+import java.math.BigDecimal
 
 interface IShoppingCartService {
 
@@ -12,7 +13,7 @@ interface IShoppingCartService {
 
     fun removeFromCart(id: Int): ShoppingCart
 
-    fun calculateTotalPrice(): Double
+    fun calculateTotalPrice(): BigDecimal
 
     fun checkout(cpf: String?, paymentOptions: PaymentOptions): Sale
 }
